requirements.txt
pandas
requests
sentence-transformers
faiss-cpu
flask
